Particle: Move advection kernels into a seaprate submodule `kernels`
