+++ parcels/__init__.py
@@ -3,3 +3,4 @@ from parcels.particle import *  # NOQA
 from parcels.field import *  # NOQA
 from parcels.kernel import *  # NOQA
 import parcels.rng as random  # NOQA
+from parcels.kernels.advection import *  # noqa
