+++ parcels/kernels/__init__.py
@@ -1,0 +1,1 @@
+from .advection import *  # noqa
